revised Gateway RPC Proxy methods
